remove `runtests` command and simplify `test` command
