+++ tests/selfhosting/test_tokenizer.py
@@ -3,17 +3,7 @@ from pathlib import Path
 from pytest import CaptureFixture
 
 from aaa.runner.runner import Runner
-from aaa.runner.test_runner import TestRunner
 from aaa.tokenizer.tokenizer import Tokenizer
-
-
-def test_tokenizer_unittests() -> None:
-    test_runner = TestRunner("examples/selfhosting")
-    exit_code = test_runner.run(
-        compile=True, binary=None, run=True, runtime_type_checks=True
-    )
-
-    assert exit_code == 0
 
 
 def tokenize_with_python(aaa_file: Path) -> str:
